requirements.txt
torch
Flask
opencv-python-headless
requests
numpy
pandas
pillow
torchvision
scikit-learn